Intermediate save before proceeding with 21st March edits to ccseq rules.
This checkpoint works upto a point but needs further development to see whether it can be made to work. Inconsistencies are: 1. some uses of id in ccseq (should be o().). 2. Some rules require presence of types and types are not preserved during rewriting (would need to add contexts and type enrichment to rewrite mode).